git: Add the ability to ignore untracked files
Because some repos are huge with a tonne of untracked files and so the
prompt renders super slow (on the order of minutes), making the prompt
useless. This option allows us to set
PROMPT_GEOMETRY_GIT_STATUS_UNTRACKED_FILES=false to not pass --others to
git ls-files, make it much faster in those cases.